More gracefully handle exiting consumer process
SIGTERM uses a global to signal that the event loop should shut down.
If we are currently processing messages, stop the event loop after a
successful loop. Otherwise a periodic timer will eventually pick up the
message. There is still the potential for a race condition between a
thread reading from RabbitMQ and the periodic timer.